feat: making MS Sharepoint Online refresh token optional (#47)
Although it is possible for Amazon AppFlow to work with the Microsoft Sharepoint Online connector only passing the `accessToken` - the CloudFormation template still requires any value for the `refreshToken`. 

This PR updates the implementation of the Microsoft Sharepoint Online Connector Profile to amend for the empty `refreshToken` so the deployment is not failing.